requirements.txt
asgiref==3.8.1
Django==3.2.20
djangorestframework==3.12.4
Faker==26.3.0
python-dateutil==2.9.0.post0
djangorestframework-simplejwt==4.8.0

